% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_als.R
\name{cp_als}
\alias{cp_als}
\title{cp_als}
\usage{
cp_als(A, R = 2, max_iter = 2000, tol = 0.001)
}
\arguments{
\item{A}{An array containing the tensor to be decomposed.}

\item{R}{A rank. Must be smaller or equal to the smallest dim of A.}

\item{max_iter}{A maximum number of iterations to try.}

\item{tol}{Tolerance for convergence criterion.
Ends algorithm if relative Frobenius norm between true and reconstructed tensor
is smaller than tol.}
}
\value{
A list containing lambda which is scaling factors and a list of matrices.
}
\description{
Computes the canonical polyadic decomposition of a tensor using
the alternate least squares algorithm.
}
\details{
See T. Kolda, B. Bader, "Tensor decomposition and applications".
SIAM Applied Mathematics and Applications 2009.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
(res <- cp_als(A, ranks = c(2, 2, 2)))

}
\seealso{
\code{\link[=cp_als_reconstruct]{cp_als_reconstruct()}} to reconstruct the tensor from the decomposed parts.
}
