% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_als_reconstruct.R
\name{cp_als_reconstruct}
\alias{cp_als_reconstruct}
\title{cp_als_reconstruct}
\usage{
cp_als_reconstruct(cp_als_list)
}
\arguments{
\item{list_cores}{A list of cores generated from \code{\link[=cp_als]{cp_als()}}}
}
\value{
The reconstructed tensor generated from the canonical polyadic decomposition.
}
\description{
Computes the approximated tensor from a canonical polyadic decomposition.
}
\details{
See T. Kolda, B. Bader, "Tensor decomposition and applications".
SIAM Applied Mathematics and Applications 2009.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
res <- cp_als(A, R = 2)
cp_als_reconstruct(res)

}
\seealso{
\code{\link[=hosvd]{hosvd()}} to create the list of cores from the higher-order SVD.
}
