% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_reconstruct.R
\name{cp_reconstruct}
\alias{cp_reconstruct}
\title{cp_reconstruct}
\usage{
cp_reconstruct(cp_als_list)
}
\arguments{
\item{cp_als_list}{A decomposition generated by \code{\link[=cp]{cp()}}.}
}
\value{
The reconstructed tensor generated from the canonical polyadic decomposition.
}
\description{
Computes the approximated tensor from a canonical polyadic decomposition.
}
\details{
See T. Kolda, B. Bader, "Tensor decomposition and applications".
SIAM Applied Mathematics and Applications 2009.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
res <- cp(A, R = 2)
cp_reconstruct(res)

}
\seealso{
\code{\link[=cp]{cp()}} to create the list of cores from the higher-order SVD.
}
