% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{draw_shape}
\alias{draw_shape}
\title{Shape of each tensor observation}
\usage{
draw_shape(x)
}
\arguments{
\item{x}{Tensor samples.}
}
\value{
An integer vector containing the dimensions of one observation.
}
\description{
Shape of each tensor observation
}
