% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtinvgauss.R
\name{dtinvgauss}
\alias{dtinvgauss}
\title{dtinvgauss}
\usage{
dtinvgauss(x, mu, skew, sigmas, kappa, log = FALSE)
}
\arguments{
\item{x}{An array of quantiles.}

\item{mu}{An array of the mean.}

\item{skew}{An array of the skew.}

\item{sigmas}{A list of the sigma covariance matrices.}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}

\item{nu}{A parameter that describes the shape of the
inverse Gaussian}
}
\value{
The log likelihood.

#' @examples
one_draw <- rtinvgauss(n = 1, mu = 2)
dtinvgauss(one_draw[\link{1}], mu = 0, skew = 1,
sigmas = list(matrix(1)), kappa = 2)
}
\description{
Computes the observed log likelihood for the tensor variate
inverse Gaussian.
}
