% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnorm.R
\name{dtnorm}
\alias{dtnorm}
\title{dtnorm}
\usage{
dtnorm(x, mu = NULL, sigmas = NULL, log = FALSE)
}
\arguments{
\item{x}{A numeric vector/array, a \code{tensor}, or a \code{tensor} object. A
sample returns one density per observation.}

\item{mu}{An array of the mean.}

\item{sigmas}{A list of covariance matrices.}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}
}
\value{
A single density for one tensor, or a numeric vector of densities
for a tensor object.
}
\description{
Computes the density function for the tensor variate normal.
}
\examples{
dtnorm(x = array(1), mu = 0, sigmas = list(matrix(1)))
dnorm(1)
dtnorm(x = array(1:12, dim = c(2, 3)), mu = array(0, dim = c(2, 3)),
       sigmas = lapply(c(2, 3), diag))
}
