% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frob_norm.R
\name{frob_norm}
\alias{frob_norm}
\title{frob_norm}
\usage{
frob_norm(A)
}
\arguments{
\item{A}{An array containing the tensor.}
}
\value{
The Frobenius norm.
}
\description{
Computes the Frobenius norm of a tensor.
}
\details{
The norm of a tensor is the square root of the sum of the
squares of all its elements.
\deqn{||\mathcal{X}|| = \sqrt{\sum_{i_{1}=1}^{I_{1}} \sum_{i_{2}=1}^{I_{2}} \dots \sum_{i_{N}=1}^{I_{N}} x_{i_{1}, i_{2}, \dots, i_{N}}^{2}}}
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
frob_norm(A)

}
