% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frob_norm_diff.R
\name{frob_norm_diff}
\alias{frob_norm_diff}
\title{frob_norm_diff}
\usage{
frob_norm_diff(A, B)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}
}
\value{
A value representing the relative Frobenius difference between A and B.
}
\description{
Computes the relative difference of Frobenius norms between two matrices.
}
\examples{
matA <- matrix(1:4, nrow = 2)
matB <- matrix(1:4, nrow = 2) + rnorm(4)
frob_norm_diff(matA, matB)

}
