% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_nprod.R
\name{\%x1\%}
\alias{\%x1\%}
\title{n-mode product (mode 1)}
\usage{
tensor \%x1\% mat
}
\arguments{
\item{tensor}{A tensor / array.}

\item{mat}{A matrix with compatible dimensions.}
}
\description{
First mode of tensor times columns of matrix.
}
