% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_nprod.R
\name{\%x2\%}
\alias{\%x2\%}
\title{n-mode product (mode 2)}
\usage{
tensor \%x2\% mat
}
\arguments{
\item{tensor}{A tensor / array.}

\item{mat}{A matrix with compatible dimensions.}
}
\description{
Second mode of tensor times columns of matrix.
}
