% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_nprod.R
\name{\%x3\%}
\alias{\%x3\%}
\title{n-mode product (mode 3)}
\usage{
tensor \%x3\% mat
}
\arguments{
\item{tensor}{A tensor / array.}

\item{mat}{A matrix with compatible dimensions.}
}
\description{
Third mode of tensor times columns of matrix.
}
