% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosvd_reconstruct.R
\name{hosvd_reconstruct}
\alias{hosvd_reconstruct}
\title{hosvd_reconstruct}
\usage{
hosvd_reconstruct(hosvd_cores)
}
\arguments{
\item{list_cores}{A list of cores generated from \code{\link[=hosvd]{hosvd()}}}
}
\value{
The reconstructed tensor generated from the higher-order SVD.
}
\description{
Computes the approximated tensor from a higher-order SVD.
}
\details{
See Lathauwer, L. and De Moor, B. (2000),
“A Multi-Linear Singular Value Decomposition,”
Society for Industrial and Applied Mathematics, 21, 1253–1278. for more details.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
res <- hosvd(A, ranks = c(2, 2, 2))
hosvd_reconstruct(res)

}
\seealso{
\code{\link[=hosvd]{hosvd()}} to create the list of cores from the higher-order SVD.
}
