% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khatri_rao.R
\name{khatri_rao}
\alias{khatri_rao}
\title{khatri_rao}
\usage{
khatri_rao(A, B)
}
\arguments{
\item{A}{A matrix of size \eqn{I \times K}}

\item{B}{A matrix of size \eqn{J \times K}}
}
\value{
A matrix of size \eqn{IJ \times K} containing the Khatri-Rao product of A and B.
}
\description{
Computes the higher-order SVD decomposition of a tensor.
}
\details{
Also known as the columnwise Kronecker product.
Computes the Kronecker of the columns of A and B.
}
\examples{
A <- matrix(1:6, nrow = 2)
B <- matrix(1:12, nrow = 4)
khatri_rao(A, B)

}
