% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_test.R
\name{mahalanobis_test}
\alias{mahalanobis_test}
\title{mahalanobis_test}
\usage{
mahalanobis_test(distances)
}
\arguments{
\item{distances}{A data frame with one column containing the vector distances
and one column containing the tensor distances.}
}
\value{
A list containing the statistic, p-value, method, and data name.
}
\description{
Computes the Kolmogorov-Smirnov test on a data frame of tensor and vector
Mahalanobis distance.
}
\examples{
s1 <- matrix(rnorm(4), nrow = 2) |> crossprod()
s2 <- matrix(rnorm(16), nrow = 4) |> crossprod()
matnorm_draws <- rtnorm(n, mu = matrix(1:8, nrow = 2), sigmas = list(s1, s2))
distances <- mahalanobis_dist(matnorm_draws)
mahalanobis_test(distances)

}
\seealso{
\code{\link[=mahalanobis_dist]{mahalanobis_dist()}} to compute the distances and
\code{\link[=plot_dist]{plot_dist()}} to create a plot of the Malahanobis distances.
}
