% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricization.R
\name{matricization}
\alias{matricization}
\title{matricization}
\usage{
matricization(array, k)
}
\arguments{
\item{array}{An array representing a tensor.}

\item{k}{A mode to matricize along.}
}
\value{
An array: the result of the nm-mode product between A and B.
}
\description{
Compute the matricization of a tensor along the kth mode.
}
\examples{
X <- array(1:24, dim = c(2, 3, 4))
matricization(X, 1)
matricization(X, 2)
matricization(X, 3)
}
