% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_est.R
\name{mle_est}
\alias{mle_est}
\title{mle_est}
\usage{
mle_est(data, max_iter = 1000, tol = 1e-06)
}
\arguments{
\item{data}{An array containing the draws, where the first mode represents each draw.}

\item{max_iter}{A max number of iterations to try to get covariance matrices that converge.}

\item{tol}{A tolerance level to define the convergence of matrices.}
}
\value{
A list containing the estimated mean array and the list of covariance matrices.
}
\description{
Estimates the mean array and covariance matrices from an array of tensor variate normal draws.
}
\examples{
all_dims <- c(2, 10, 10, 2)
fourth_mu = array(rnorm(400), dim = all_dims)
fourth_sigma = lapply(seq_along(all_dims), function(k) diag(all_dims[k]))
fourth_tensor <- rtnorm(n = 1e3, mu = fourth_mu, sigmas = fourth_sigma)
est_fourth <- mle_est(fourth_tensor)
(fourth_mu - est_fourth$mu)^2 |> mean()
fourth_scaled <- lapply(fourth_sigma, function(S) S * (nrow(S) / sum(diag(S))))
(mse_each <- mapply(function(est, true) mean((est - true)^2), est_fourth$sigmas, fourth_scaled))
}
