% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{n_draws}
\alias{n_draws}
\title{Number of tensor observations}
\usage{
n_draws(x)
}
\arguments{
\item{x}{Tensor samples.}
}
\value{
A single integer giving the number of observations.
}
\description{
Number of tensor observations
}
