% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(
  x12,
  x22,
  x11,
  x21,
  mu,
  sigmas,
  model = c("normal", "skewt", "vargamma", "invgauss", "genhyper"),
  skew = NULL,
  nu = NULL,
  scale = NULL,
  kappa = NULL,
  lambda = NULL,
  omega = NULL,
  contour_bins = NULL,
  interpolate = TRUE
)
}
\arguments{
\item{x12}{A numeric vector of fixed values for \code{X[1, 2]}.}

\item{x22}{A numeric vector of fixed values for \code{X[2, 2]}.}

\item{x11}{A numeric vector of values to evaluate for \code{X[1, 1]}.}

\item{x21}{A numeric vector of values to evaluate for \code{X[2, 1]}.}

\item{mu}{A 2 by 2 location matrix.}

\item{sigmas}{A list of covariance matrices.}

\item{model}{The model name. One of \code{"normal"}, \code{"skewt"}, \code{"vargamma"},
\code{"invgauss"}, or \code{"genhyper"}.}

\item{skew}{A 2 by 2 skewness matrix. Required for the skewed models.}

\item{nu}{Degrees of freedom. Required when \code{model = "skewt"}.}

\item{scale}{Scale parameter. Required when \code{model = "vargamma"}.}

\item{kappa}{Kappa parameter. Required when \code{model = "invgauss"}.}

\item{lambda}{Lambda parameter. Required when \code{model = "genhyper"}.}

\item{omega}{Omega parameter. Required when \code{model = "genhyper"}.}

\item{contour_bins}{Number of contour bins to draw.}

\item{interpolate}{Logical. If TRUE, interpolate the raster display.}
}
\value{
A ggplot object.
}
\description{
Plots two-dimensional density slices for a 2 by 2 matrix-variate model.
}
\examples{
mu <- matrix(c(0, 0, 0, 0), nrow = 2)
sigmas <- list(diag(2), diag(2))
plot_density(
  x12 = seq(-2, 2, by = 1),
  x22 = seq(-2, 2, by = 1),
  x11 = seq(-4, 4, length.out = 25),
  x21 = seq(-4, 4, length.out = 25),
  mu = mu,
  sigmas = sigmas,
  model = "normal"
)

}
