% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{print.tensor_samples}
\alias{print.tensor_samples}
\title{Print tensor samples}
\usage{
\method{print}{tensor_samples}(
  x,
  ...,
  n = getOption("tensormodels.print_draws", 3L),
  entries = getOption("tensormodels.print_entries", 6L),
  digits = getOption("digits", 7L)
)
}
\arguments{
\item{x}{Tensor samples.}

\item{...}{Additional arguments, currently unused.}

\item{n}{Number of observations to include in the preview.}

\item{entries}{Number of indexed entries to show per observation.}

\item{digits}{Number of significant digits used to print numeric values.}
}
\value{
\code{x}, invisibly.
}
\description{
Print tensor samples
}
