% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{pull_draw}
\alias{pull_draw}
\title{Extract one tensor draw}
\usage{
pull_draw(x, i)
}
\arguments{
\item{x}{Tensor samples.}

\item{i}{A single numeric or character location identifying a draw.}
}
\value{
A \code{tensor} containing one tensor observation.
}
\description{
Extract one tensor draw
}
