% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtnorm.R
\name{rtnorm}
\alias{rtnorm}
\title{rtnorm}
\usage{
rtnorm(n = 50, mu = 0, listSigmas = NULL)
}
\arguments{
\item{n}{An integer stating the sample size.}

\item{mu}{An array containing the mean values of and dims of each draw.}

\item{listSigmas}{A list of covariance matrices. Defaults to the identity.}
}
\value{
An array containing n draws of the tensor variate normal.
}
\description{
Simulate random draws from the tensor variate normal distribution
}
\examples{
univarNorm <- rtnorm(n = 10000, mu = -2)
mean(univarNorm)
sd(univarNorm)
matrixVarNorm <- matrixVarNorm <- rtnorm(n = 10000, mu = matrix(1:6, nrow = 2, ncol = 3),
listSigmas = list(diag(2), diag(3)))
}
