% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtvargamma.R
\name{rtvargamma}
\alias{rtvargamma}
\title{rtvargamma}
\usage{
rtvargamma(n, mu = 0, sigmas = 1, skew = 1, scale = 2)
}
\arguments{
\item{n}{An integer stating the sample size.}

\item{mu}{An array containing the mean values of and dims of each draw.}

\item{sigmas}{A list of covariance matrices. Defaults to the identity.}

\item{skew}{An array that determines how skewed the distribution is.}

\item{scale}{Scale parameter for gamma.}
}
\value{
An array containing n draws of the tensor variate variance gamma distribution.
}
\description{
Simulate random draws from the tensor variate variance gamma distribution.
}
\examples{
univar_vargam <- rtskewt(n = 1e3, mu = 0, skew = 0, scale = 2)
mean(univar_vargam)
sd(univar_vargam)
}
