% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{slice_draws}
\alias{slice_draws}
\title{Select tensor draws}
\usage{
slice_draws(x, i)
}
\arguments{
\item{x}{Tensor samples.}

\item{i}{Locations identifying draws to retain.}
}
\value{
A \code{tensor_samples} object containing the selected observations.
}
\description{
Select tensor draws
}
