% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_tensor.R
\name{str.tensor}
\alias{str.tensor}
\title{Structure of a tensor.}
\usage{
\method{str}{tensor}(object, ...)
}
\arguments{
\item{A}{A tensor object (S3 wrapper around an array).}
}
\value{
The structure of a tensor.
}
\description{
Structure of a tensor.
}
