% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tensor.R
\name{[<-.tensor}
\alias{[<-.tensor}
\title{Assign a tensor.}
\usage{
\method{[}{tensor}(x, ..., value)
}
\arguments{
\item{...}{Indices.}

\item{value}{Tensor that will be assigned to A.}

\item{A}{A tensor object (S3 wrapper around an array).}
}
\value{
A tensor that was assigned.
}
\description{
Assign a tensor.
}
\examples{
A <- tensor(1:24, dim = c(2, 3, 4))
A[1, ] <- array(1:12, dim = c(3, 4))
}
