% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{Construct a tensor}
\usage{
tensor(x, dim = NULL, dimnames = NULL)
}
\arguments{
\item{x}{An array, or the data used to construct one.}

\item{dim}{An optional integer vector giving the dimensions of the tensor.
If \code{NULL}, \code{x} must already be an array.}

\item{dimnames}{An optional list of dimension names used when \code{dim} is
supplied.}
}
\value{
An object of class \code{tensor} backed by \code{x}.
}
\description{
\code{tensor()} wraps a base array as a tensor, or constructs an array from raw
data when \code{dim} is supplied. Every array dimension represents a tensor mode.
}
\examples{
A <- tensor(array(1:24, dim = c(2, 3, 4)))
B <- tensor(1:24, dim = c(2, 3, 4))
A + A
2 * A

}
