% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tensor_prod}
\alias{tensor_prod}
\title{tensor product}
\usage{
tensor_prod(A, B, simplify = TRUE)
}
\arguments{
\item{simplify}{A bool saying whether to simplify the output. Defaults to TRUE.}

\item{tensor_A}{An array representing the tensor A.}

\item{tensor_B}{An array representing tensor B.}
}
\value{
An array: the result of the tensor product between A and B.
}
\description{
Compute the tensor product of a tensor with a tensor
}
\examples{
A <- matrix(c(1, 2, 3, 4), nrow = 2)
x <- matrix(c(5, 6), nrow = 2)
tensor_prod(A, x)
}
