% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_reconstruct.R
\name{tt_reconstruct}
\alias{tt_reconstruct}
\title{tt_reconstruct}
\usage{
tt_reconstruct(list_cores)
}
\arguments{
\item{list_cores}{A list of cores generated from \code{\link[=tt]{tt()}}}
}
\value{
The reconstructed tensor generated from the tensor-train decomposition.
}
\description{
Computes the approximated tensor from a tensor-train decomposition.
}
\details{
See I. V. Oseledets, (2011). Tensor-Train Decomposition.
SIAM J. SCI. COMPUT. for more details.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
res <- tt_svd(A, ranks = c(2, 2, 2))
tt_svd_reconstruct(res)

}
\seealso{
\code{\link[=tt]{tt()}} to create the list of cores from the tensor-train decomposition.
}
