% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_svd.R
\name{tt_svd}
\alias{tt_svd}
\title{tt_svd}
\usage{
tt_svd(A, ranks = NULL, epsilon = NULL)
}
\arguments{
\item{A}{An array containing the tensor to be decomposed.}

\item{ranks}{A list of ranks used for the approximation.
If provided, accuracy will be ignored.}

\item{accuracy}{Determines the accuracy of the approximation.
If ranks are not provided, will compute them based on \eqn{\delta}-truncation}
}
\value{
A list of the tensor-train cores used to reconstruct the array.
}
\description{
Computes the tensor-train decomposition of a tensor using the SVD.
}
\details{
See I. V. Oseledets, (2011). Tensor-Train Decomposition. SIAM J. SCI. COMPUT. for more details.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
res <- tt_svd(A, ranks = c(2, 2, 2))

}
\seealso{
\code{\link[=tt_svd_reconstruct]{tt_svd_reconstruct()}} to reconstruct the tensor from the decomposed parts.
}
