% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucker.R
\name{tucker}
\alias{tucker}
\title{tucker}
\usage{
tucker(A, ranks = ranks, max_iter = 2000, tol = 0.001)
}
\arguments{
\item{A}{An array containing the tensor to be decomposed.}

\item{max_iter}{A maximum number of iterations to try.}

\item{tol}{Tolerance for convergence criterion.
Ends algorithm if relative Frobenius norm between true and reconstructed tensor
is smaller than tol.}

\item{R}{A list of ranks of the same size as order of A. E
Each rank must be smaller or equal to corresponding dim.}
}
\value{
A list containing a core tensor and a list of matrices.
}
\description{
Computes the canonical polyadic decomposition of a tensor using
the alternate least squares algorithm.
}
\details{
See T. Kolda, B. Bader, "Tensor decomposition and applications".
SIAM Applied Mathematics and Applications 2009.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
(res <- tucker(A, ranks = c(2, 2, 2)))

}
\seealso{
\code{\link[=tucker_reconstruct]{tucker_reconstruct()}} to reconstruct the tensor from the decomposed parts.
}
