% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucker_reconstruct.R
\name{tucker_reconstruct}
\alias{tucker_reconstruct}
\title{tucker_reconstruct}
\usage{
tucker_reconstruct(tucker_cores)
}
\arguments{
\item{list_cores}{A list of cores generated from \code{\link[=tucker]{tucker()}}}
}
\value{
The reconstructed tensor generated from the Tucker decomposition.
}
\description{
Computes the approximated tensor from a Tucker decomposition
}
\details{
See T. Kolda, B. Bader, "Tensor decomposition and applications".
SIAM Applied Mathematics and Applications 2009. for more details.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
res <- tucker(A, ranks = c(2, 2, 2))
tucker_reconstruct(res)

}
\seealso{
\code{\link[=tucker]{tucker()}} to create the list of cores from the Tucker decomposition.
}
