% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tensor.R
\name{as_tensor}
\alias{as_tensor}
\title{Coerce an array to a tensor.}
\usage{
as_tensor(x)
}
\arguments{
\item{x}{An array.}
}
\value{
An object of class \code{tensor}.
}
\description{
Coerce an array to a tensor.
}
