% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{draw_shape}
\alias{draw_shape}
\title{Shape of each tensor draw}
\usage{
draw_shape(x)
}
\arguments{
\item{x}{A tensor object.}
}
\value{
Dimensions of one draw.
}
\description{
Shape of each tensor draw
}
