% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtgenhyper.R
\name{dtgenhyper}
\alias{dtgenhyper}
\title{dtgenhyper
Computes the density function for the tensor variate generalized hyperbolic.}
\usage{
dtgenhyper(x, mu, skew, sigmas, lambda, omega, log = FALSE)
}
\arguments{
\item{x}{An array of quantiles.}

\item{mu}{An array of the mean.}

\item{skew}{An array of the skew.}

\item{sigmas}{A list of the sigma covariance matrices.}

\item{lambda}{A parameter that describes the shape of the
generalized inverse Gaussian.}

\item{omega}{A parameter that describes the shape and scale of the
generalized inverse Gaussian.}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}
}
\value{
The log likelihood.
}
\description{
dtgenhyper
Computes the density function for the tensor variate generalized hyperbolic.
}
\examples{
dtgenhyper(array(1), mu = 0, skew = array(1),
           sigmas = list(matrix(1)), lambda = 2, omega = 2)
}
