% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtinvgauss.R
\name{dtinvgauss}
\alias{dtinvgauss}
\title{dtinvgauss}
\usage{
dtinvgauss(x, mu, skew, sigmas, kappa, log = FALSE)
}
\arguments{
\item{x}{An array of quantiles or a list of arrays.}

\item{mu}{An array of the mean.}

\item{skew}{An array of the skew.}

\item{sigmas}{A list of the sigma covariance matrices.}

\item{kappa}{A parameter that describes the shape of the
inverse Gaussian}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}
}
\value{
The log likelihood.
}
\description{
Computes the observed log likelihood for the tensor variate
inverse Gaussian.
}
\examples{
dtinvgauss(array(1), mu = 0, skew = array(1), sigmas = list(matrix(1)), kappa = 2)
}
