% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnorm.R
\name{dtnorm}
\alias{dtnorm}
\title{dtnorm}
\usage{
dtnorm(x, mu, sigmas, log = FALSE)
}
\arguments{
\item{x}{An array of quantiles.}

\item{mu}{An array of the mean.}

\item{sigmas}{A list of covariance matrices.}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}
}
\value{
The log likelihood.
}
\description{
Computes the density function for the tensor variate normal.
}
\examples{
one_draw <- rnorm(n = 1)
dtnorm(one_draw, mu = 0, sigmas = list(matrix(1)))
dnorm(one_draw)
}
