% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtskewt.R
\name{dtskewt}
\alias{dtskewt}
\title{dtskewt}
\usage{
dtskewt(x, mu, skew, sigmas, nu, log = FALSE)
}
\arguments{
\item{x}{An array of quantiles.}

\item{mu}{An array of the mean.}

\item{skew}{An array of the skew.}

\item{sigmas}{A list of the sigma covariance matrices.}

\item{nu}{A parameter that describes the shape and rate of the
inverse gamma.}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}
}
\value{
The log likelihood.
}
\description{
Computes the density function for the tensor variate skewed-t.
}
\examples{
one_draw <- rtskewt(n = 1, mu = 2)
dtskewt(one_draw[[1]], mu = 0, skew = 1, sigmas = list(matrix(1)), nu = 2)
}
