% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtvargamma.R
\name{dtvargamma}
\alias{dtvargamma}
\title{dtvargamma}
\usage{
dtvargamma(x, mu, skew, sigmas, scale, log = FALSE)
}
\arguments{
\item{x}{An array of quantiles or a list of arrays.}

\item{mu}{An array of the mean.}

\item{skew}{An array of the skew.}

\item{sigmas}{A list of the sigma covariance matrices.}

\item{scale}{A parameter that describes the shape and rate of the gamma.}

\item{log}{Defaults to FALSE. If TRUE, return the log of the density.}
}
\value{
The log likelihood.
}
\description{
Computes the density function for the tensor variate gamma.
}
\examples{
dtvargamma(array(1), mu = 0, skew = array(1), sigmas = list(matrix(1)), scale = 4)
}
