% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_est.R
\name{em_est}
\alias{em_est}
\title{em_est}
\usage{
em_est(draws, max_iter = 1000, tol = 1e-06, quiet = TRUE, model)
}
\arguments{
\item{draws}{An array containing the draws, where the first mode represents
each draw.}

\item{max_iter}{A max number of iterations to try to get covariance matrices
that converge.}

\item{tol}{A tolerance level to define the convergence of matrices.}

\item{method}{Specify the distribution of draws. Must be one of skewt,
vargamma, invgauss, or gen hyper}
}
\value{
A list containing the estimated mean array and the list of covariance
matrices.
}
\description{
Estimates the mean array and covariance matrices from an array of tensor
variate normal draws using an expectation-maximization algorithm.
}
