% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosvd.R
\name{hosvd}
\alias{hosvd}
\title{hosvd}
\usage{
hosvd(A, ranks = NULL)
}
\arguments{
\item{A}{An array containing the tensor to be decomposed.}

\item{ranks}{A list of ranks used for the approximation. Must be same length as order of A.}
}
\value{
A list containing a core tensor and a list of matrices.
}
\description{
Computes the higher-order SVD decomposition of a tensor.
}
\details{
See Lathauwer, L. and De Moor, B. (2000),
“A Multi-Linear Singular Value Decomposition,”
Society for Industrial and Applied Mathematics, 21, 1253–1278. for more details.
}
\examples{
A <- array(1:24, dim = c(2, 3, 4))
(res <- hosvd(A, ranks = c(2, 2, 2)))

}
\seealso{
\code{\link[=hosvd_reconstruct]{hosvd_reconstruct()}} to reconstruct the tensor from the decomposed parts.
}
