% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_dist.R
\name{mahalanobis_dist}
\alias{mahalanobis_dist}
\title{mahalanobis_dist}
\usage{
mahalanobis_dist(data)
}
\arguments{
\item{data}{A list of arrays containing the independent, identically distributed draws.}
}
\value{
A vector of distances.
}
\description{
Computes the Mahalanobis distances for data using the MLE.
}
\examples{
s1 <- matrix(rnorm(4), nrow = 2) |> crossprod()
s2 <- matrix(rnorm(16), nrow = 4) |> crossprod()
matnorm_draws <- rtnorm(n = 10, mu = matrix(1:8, nrow = 2), sigmas = list(s1, s2))
(mahalanobis_dist(matnorm_draws))

}
