% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_mode_prod.R, R/nm_prod.R
\name{n_mode_prod}
\alias{n_mode_prod}
\title{n-mode product}
\usage{
n_mode_prod(tensorA, tensorB, n, m)

n_mode_prod(tensorA, tensorB, n, m)
}
\arguments{
\item{tensorA}{An array representing the tensor A.}

\item{tensorB}{An array representing tensor B.}

\item{n}{An integer specifying which mode of tensor A to multiply across.}

\item{m}{An integer specifying which mode of tensor B to multiply across.}

\item{tensor}{An array representing the tensor.}

\item{mat}{A matrix of size \eqn{m \times n} to multiply with.}
}
\value{
An array: the result of the n-mode product.

An array: the result of the nm-mode product between A and B.
}
\description{
Compute the n-mode product of a tensor with a matrix.

Compute the nm-mode product of a tensor with a tensor
}
\examples{
a <- array(1:3, dim = c(3, 1, 1))
b <- matrix(4:9, nrow = 2, ncol = 3)
n_mode_prod(a, b, 1)
a <- array(1:3, dim = c(3, 1, 1))
b <- matrix(4:9, nrow = 2, ncol = 3)
n_mode_prod(a, b, 1)
}
