% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_prod}
\alias{n_prod}
\title{n-mode product}
\usage{
n_prod(tensor, mat, n)
}
\arguments{
\item{tensor}{An array representing the tensor.}

\item{mat}{A matrix of size \eqn{m \times n} to multiply with.}

\item{n}{An integer specifying the mode to multiply across.}
}
\value{
An array: the result of the n-mode product.
}
\description{
Compute the n-mode product of a tensor with a matrix.
}
\details{
The $n$-mode matrix product of a tensor
$\tensor{A} \in \field{k}^{I_1 \times I_2 \times ... \times I_n}$
with a matrix $\mat{U} \in \field{k}^{J \times I_n}$ is
$$(\tensor{A} \times_n \mat{U}) \in
\field{k}^{I_1 \times ... \times I_{n-1} \times J \times I_{n+1} \times ... \times I_N}$$
with entries $$(\tensor{A} \times_n \mat{U})\emph{{i_1, ..., i}{n-1}, j, i_{n+1}, ... I_N} =
\sum_{i_n = 1}^{I_n} a_{i_1, i_2, \dots, i_N} u_{j, i_n}.$$
The tensor and matrix share one mode in common, denoted here as $I_n$.
This operation is also called the tensor times matrix product.
}
\examples{
a <- array(1:3, dim = c(3, 1, 1))
x <- matrix(4:9, nrow = 2, ncol = 3)
n_prod(a, x, 1)
}
