% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/n_prod.R
\name{n_prod}
\alias{n_prod}
\title{n-mode product}
\usage{
n_prod(tensor, n, mat)

n_prod(tensor, n, mat)
}
\arguments{
\item{tensor}{A tensor object or an array.}

\item{n}{The mode within each tensor draw.}

\item{mat}{A compatible matrix.}
}
\value{
An array: the result of the n-mode product.

A tensor when \code{tensor} is a tensor object, otherwise an array.
}
\description{
Compute the n-mode product of a tensor with a matrix.
}
\details{
The \eqn{n}-mode matrix product of a tensor
\eqn{\tensor{A} \in \field{k}^{I_1 \times I_2 \times ... \times I_n}}
with a matrix \eqn{\mat{U} \in \field{k}^{J \times I_n}} is
\deqn{(\tensor{A} \times_n \mat{U}) \in
         \field{k}^{I_1 \times ... \times I_{n-1} \times J \times I_{n+1} \times ... \times I_N}}
with entries \deqn{(\tensor{A} \times_n \mat{U})_{i_1, ..., i_{n-1}, j, i_{n+1}, ... I_N} =
         \sum_{i_n = 1}^{I_n} a_{i_1, i_2, \dots, i_N} u_{j, i_n}.}
The tensor and matrix share one mode in common, denoted here as \en{I_n}.
This operation is also called the tensor times matrix product.
}
\examples{
a <- array(1:3, dim = c(3, 1, 1))
x <- matrix(4:9, nrow = 2, ncol = 3)
n_prod(a, 1, x)
}
