% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_prod}
\alias{n_prod}
\title{n-mode product}
\usage{
n_prod(tensor, mat, n)
}
\arguments{
\item{tensor}{An array representing the tensor.}

\item{mat}{A matrix of size \eqn{m \times n} to multiply with.}

\item{n}{An integer specifying the mode to multiply across.}
}
\value{
An array: the result of the n-mode product.
}
\description{
Compute the n-mode product of a tensor with a matrix.
}
\details{
Test
}
\examples{
a <- array(1:3, dim = c(3, 1, 1))
x <- matrix(4:9, nrow = 2, ncol = 3)
n_prod(a, x, 1)
}
