% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prod.R
\name{nm_prod}
\alias{nm_prod}
\title{nm-mode product}
\usage{
nm_prod(tensor_A, tensor_B, n, m)
}
\arguments{
\item{tensor_A}{An array representing the tensor A.}

\item{tensor_B}{An array representing tensor B.}

\item{n}{An integer specifying which mode of tensor A to multiply across.}

\item{m}{An integer specifying which mode of tensor B to multiply across.}
}
\value{
An array: the result of the nm-mode product between A and B.
}
\description{
Compute the nm-mode product of a tensor with a tensor
}
\examples{
A <- matrix(c(1, 2, 3, 4), nrow = 2)
b <- matrix(c(5, 6), nrow = 2)
nm_prod(A, b, 1, 1)
}
