% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{plot_dist}
\usage{
plot_dist(distances)
}
\arguments{
\item{distances}{A data frame with one column containing the vector distances and
one column containing the tensor distances.}
}
\value{
A plot of the vector and tensor Mahalanobis distances.
}
\description{
Plots the vector and tensor Mahalanobis distance.
}
\examples{
s1 <- matrix(rnorm(4), nrow = 2) |> crossprod()
s2 <- matrix(rnorm(16), nrow = 4) |> crossprod()
matnorm_draws <- rtnorm(n, mu = matrix(1:8, nrow = 2), sigmas = list(s1, s2))
distances <- mahalanobis_dist(matnorm_draws)
plot_dist(distances)

}
\seealso{
\code{\link[=mahalanobis_dist]{mahalanobis_dist()}} to compute the distances.
}
