% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tensor.R
\name{print.tensor}
\alias{print.tensor}
\title{Prints a tensor.}
\usage{
\method{print}{tensor}(
  x,
  ...,
  max_slices = getOption("tensor.max_slices", 3),
  edge = getOption("tensor.edge", 6),
  stats = getOption("tensor.stats", TRUE)
)
}
\arguments{
\item{A}{A tensor object (S3 wrapper around an array).}
}
\value{
Prints the tensor.
}
\description{
Prints a tensor.
}
