% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tensor.R, R/tensor_samples.R
\name{print.tensor}
\alias{print.tensor}
\title{Prints a tensor.}
\usage{
\method{print}{tensor}(
  x,
  ...,
  n = getOption("tensormodels.print_draws", 3L),
  entries = getOption("tensormodels.print_entries", 6L),
  digits = getOption("digits", 7L)
)

\method{print}{tensor}(
  x,
  ...,
  n = getOption("tensormodels.print_draws", 3L),
  entries = getOption("tensormodels.print_entries", 6L),
  digits = getOption("digits", 7L)
)
}
\arguments{
\item{x}{A tensor object.}

\item{...}{Unused.}

\item{n}{Number of draws to preview.}

\item{entries}{Number of entries per draw to preview.}

\item{digits}{Significant digits for the preview.}

\item{A}{A tensor object (S3 wrapper around an array).}
}
\value{
Prints the tensor.

\code{x}, invisibly.
}
\description{
Prints a tensor.

Print tensor-valued observations
}
