% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{print.tensor}
\alias{print.tensor}
\title{Print tensor-valued observations}
\usage{
\method{print}{tensor}(
  x,
  ...,
  n = getOption("tensormodels.print_draws", 3L),
  entries = getOption("tensormodels.print_entries", 6L),
  digits = getOption("digits", 7L)
)
}
\arguments{
\item{x}{A tensor object.}

\item{...}{Unused.}

\item{n}{Number of draws to preview.}

\item{entries}{Number of entries per draw to preview.}

\item{digits}{Significant digits for the preview.}
}
\value{
\code{x}, invisibly.
}
\description{
Print tensor-valued observations
}
