% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{pull_draw}
\alias{pull_draw}
\title{Extract one tensor draw}
\usage{
pull_draw(x, i)
}
\arguments{
\item{x}{A tensor object.}

\item{i}{Draw location.}
}
\value{
A one-draw tensor.
}
\description{
Extract one tensor draw
}
