% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgenhyper.R
\name{rtgenhyper}
\alias{rtgenhyper}
\title{rtgenhyper}
\usage{
rtgenhyper(
  n,
  mu = 0,
  skew = 1,
  sigmas = list(matrix(1)),
  lambda = 2,
  omega = 2
)
}
\arguments{
\item{n}{An integer stating the sample size.}

\item{mu}{An array containing the mean values of and dims of each draw.}

\item{skew}{An array that determines how skewed the distribution is.}

\item{sigmas}{A list of covariance matrices. Defaults to the identity.}

\item{lambda}{Shape/scale parameter for generalized inverse Gaussian distribution.}

\item{omega}{Shape parameter for generalized inverse Gaussian distribution.}
}
\value{
A \code{tensor} object containing n draws of the tensor
generalized hyperbolic distribution, with observations stored on the
first axis.
}
\description{
Simulate random draws from the tensor variate generalized hyperbolic distribution.
}
\examples{
univar_genhyper <- rtgenhyper(n = 1e3, mu = 0)
mean(univar_genhyper)
sd(univar_genhyper)
}
