% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgenhyper.R
\name{rtgenhyper}
\alias{rtgenhyper}
\title{rtgenhyper}
\usage{
rtgenhyper(n, mu = 0, sigmas = 1, skew = 1, omega = 2, lambda = 2)
}
\arguments{
\item{n}{An integer stating the sample size.}

\item{mu}{An array containing the mean values of and dims of each draw.}

\item{sigmas}{A list of covariance matrices. Defaults to the identity.}

\item{skew}{An array that determines how skewed the distribution is.}

\item{omega}{Shape parameter for generalized inverse Gaussian distribution.}

\item{lambda}{Shape/scale parameter for generalized inverse Gaussian distribution.}
}
\value{
An array containing n draws of the tensor generalized hyperbolic distribution.
}
\description{
Simulate random draws from the tensor variate generalized hyperbolic distribution.
}
\examples{
univar_skewt <- rtskewt(n = 1e3, mu = 0, skew = 0, v = 4)
mean(univar_skewt)
sd(univar_skewt)
}
