% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtnorm.R
\name{rtnorm}
\alias{rtnorm}
\title{rtnorm}
\usage{
rtnorm(n, mu = 0, sigmas = 1)
}
\arguments{
\item{n}{An integer stating the sample size.}

\item{mu}{An array containing the mean values of and dims of each draw.}

\item{sigmas}{A list of covariance matrices. Defaults to the identity.}
}
\value{
An array containing n draws of the tensor variate normal distribution.
}
\description{
Simulate random draws from the tensor variate normal distribution.
\deqn{\frac{(x-\mu)^2}{2\sigma^2}}
}
\examples{
univar_norm <- rtnorm(n = 10000, mu = -2)
mean(univar_norm)
sd(univar_norm)
matrix_var_norm <- matrix_var_norm <- rtnorm(n = 10000, mu = matrix(1:6, nrow = 2, ncol = 3))
}
