% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtskewt.R
\name{rtskewt}
\alias{rtskewt}
\title{rtskewt}
\usage{
rtskewt(n, mu = 0, sigmas = list(matrix(1)), skew = 1, nu = 4)
}
\arguments{
\item{n}{An integer stating the sample size.}

\item{mu}{An array containing the mean values of and dims of each draw.}

\item{sigmas}{A list of covariance matrices. Defaults to the identity.}

\item{skew}{An array that determines how skewed the distribution is.}

\item{v}{Degrees of freedom}
}
\value{
An array containing n draws of the tensor variate skewed t distribution.
}
\description{
Simulate random draws from the tensor variate skewed t distribution.
}
\examples{
univar_skewt <- rtskewt(n = 1e3, mu = 0)
mean(univar_skewt)
sd(univar_skewt)
}
