% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_samples.R
\name{slice_draws}
\alias{slice_draws}
\title{Select tensor draws}
\usage{
slice_draws(x, i)
}
\arguments{
\item{x}{A tensor object.}

\item{i}{Draw locations.}
}
\value{
A tensor containing selected draws.
}
\description{
Select tensor draws
}
