% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_tensor.R
\name{str.tensor}
\alias{str.tensor}
\title{Structure of a tensor}
\usage{
\method{str}{tensor}(object, ...)
}
\arguments{
\item{object}{A tensor object.}

\item{...}{Unused.}
}
\value{
A compact description.
}
\description{
Structure of a tensor
}
