% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_tensor.R
\name{[.tensor}
\alias{[.tensor}
\title{Subset a tensor.}
\usage{
\method{[}{tensor}(x, ..., drop = TRUE)
}
\arguments{
\item{...}{Indices.}

\item{drop}{Logical; passed to base \code{[} (defaults to FALSE).}

\item{A}{A tensor object (S3 wrapper around an array).}
}
\value{
A tensor that was subset.
}
\description{
Subset a tensor.
}
\examples{
A <- tensor(1:24, dim = c(2, 3, 4))
A[1, ]
}
