% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tensor.R
\name{[<-.tensor}
\alias{[<-.tensor}
\title{Assign into a tensor}
\usage{
\method{[}{tensor}(x, ...) <- value
}
\arguments{
\item{x}{A tensor object.}

\item{...}{Array indices.}

\item{value}{A replacement value, including a tensor or ordinary array.}
}
\value{
The modified tensor.
}
\description{
Assignment follows ordinary array rules. For a one-draw tensor, the draw
axis is implicit and is inserted automatically.
}
