% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{tensor}
\usage{
tensor(data, dims)
}
\arguments{
\item{data}{A vector containing data for the tensor.}

\item{dims}{A vector of dimensions for the resulting tensor.}
}
\value{
A tensor with values from data and dimensions from dims.
}
\description{
Creates a tensor object.
}
\examples{
(A <- tensor(1:24, dim = c(2, 3, 4)))

}
