% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_mle.R
\name{tensor_mle}
\alias{tensor_mle}
\title{tensor_mle}
\usage{
tensor_mle(
  draws,
  max_iter = 1000,
  tol = 1e-04,
  quiet = TRUE,
  model,
  restrict = NULL
)
}
\arguments{
\item{draws}{A \code{tensor} object containing IID observations.}

\item{max_iter}{A max number of iterations to try to get covariance matrices
that converge.}

\item{tol}{A tolerance level to define the convergence of matrices.}

\item{model}{Specify the distribution of draws. Must be one of normal, skewt,
vargamma, invgauss, or genhyper}
}
\value{
A list containing the estimated mean array and the list of covariance
matrices.
}
\description{
Estimates the mean array and covariance matrices from tensor-valued
observations. If tensor variate normal, uses the flop-flop algorithm. Otherwise,
uses the expectation-maximization algorithm.
}
