% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_mle.R
\name{tensor_mle}
\alias{tensor_mle}
\title{tensor_mle}
\usage{
tensor_mle(
  draws,
  max_iter = 1000,
  tol = 0.01,
  quiet = TRUE,
  model,
  restrict = NULL
)
}
\arguments{
\item{draws}{An array containing the draws, where the first mode represents
each draw.}

\item{max_iter}{A max number of iterations to try to get covariance matrices
that converge.}

\item{tol}{A tolerance level to define the convergence of matrices.}

\item{method}{Specify the distribution of draws. Must be one of normal, skewt,
vargamma, invgauss, or gen hyper}
}
\value{
A list containing the estimated mean array and the list of covariance
matrices.
}
\description{
Estimates the mean array and covariance matrices from an array of tensor
variate draws. If tensor variate normal, uses the flop-flop algorithm. Otherwise,
uses the expectation-maximization algorithm.
}
